Remove CharcoalIdentifiableOverlayView out
